Main: Exclude SDCC-required functions from dce
This commit introduces a list of functions that should not be optimized out by the stm8dce tool, as they are required by the SDCC standard library.

A hardcoded list is necessary because the standard library is not compiled together with the source files and is thus out-of-reach for the stm8dce tool.

Currently, the list only includes `_getchar` and `_putchar`, which are required by `puts.c` and `gets.c` in the SDCC standard library. According to the SDCC manual, `_getchar` and `_putchar` are the only functions that must be provided by the user. However, please let me know if I am mistaken about this.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>