Improved detection of constants being read
This commit addresses the issue of constants being optimized away due to inadequate label read detection.

Previously, the tool only checked if a constant was used by verifying if it was loaded with an `ld` or `ldw` instruction. This method was insufficient because in some scenarios, the constant's address is pushed onto the stack rather than being loaded. This pushing of constants was more common in older versions of SDCC but can still occur in the latest version (4.4.1).

This update expands the detection to analyze **all** instructions capable of reading from a long memory address. The instructions now considered are:

    ld, ldf, ldw, mov, adc, add, and, bcp, cp, or, sub, sbc, xor, addw, subw, cpw,
    btjt, btjf, push, call, callf, jp, jpf, int

The following resources were referenced to determine this list:
  - PM0044 STM8 CPU Programming Manual
  - SDCC Source Code:
    - sdcc/sdas/asstm8/stm8mch.c
    - sdcc/sdas/asstm8/stm8pst.c
    - sdcc/sdas/asstm8/stm8.h

Additionally, the algorithm for detecting labels has been improved. It no longer relies heavily on SDCC-generated "quirks" (e.g., read labels always including an offset, even if it's 0). This enhancement aims to ensure better future compatibility should SDCC change its label usage.

This commit should resolve the issues mentioned in:
- https://github.com/CTXz/STM8-DCE/issues/3
- https://github.com/maxgerhardt/pio-stm8-dce/commit/6dcdcde63420e9a74418933b8de94c756cdd9a1c