Add Support for .rel and .lib File Analysis
This commit introduces support for relocatable object files, extending stm8dce's capabilities to handle .rel and .lib files in addition to .asm files.

When a .lib or .rel file is added (note that .lib files are just collections of mutiple .rel files), stm8dce:

- First: Analyzes which modules are used by the .asm files
- Then: Evaluates if the module references a symbol from the .asm file.
- Finally: Functions/constants assigned to the referenced symbols will not be eliminated.

Note that this commit does NOT perform dce on rel and lib files themselves. It only uses them to gather valuable information for the asm optimization.

As a pratical example, consider SDCC's print implementation. If `print()` is called within the project, `putchar()` is expected to be defined. By adding the standard library (`stm8lib.lib`) to stm8dce, it detects that `print()` is an external call to the `vprint` module in `stm8lib.lib`. It then realizes that the `vprint` module references `putchar()` and ensures `putchar()` is not eliminated.

> Small Side Note: `print(str)` without format specifiers is a macro for `puts`. Therefore, `print()` may also use the `puts` module instead of `vprint` if only a string is printed.

For more detailed information about .rel files, please refer to:
  https://sourceforge.net/p/sdcc/code/HEAD/tree/trunk/sdcc/sdas/doc/format.txt

This commit also replaces the solution provided in d062c608d15a82608d7e7018cef83f8ebebeef54, with the added advantage that `_putchar` and `_getchar` will be optimized if no function requires them.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>