Add support for external entry points
This commit adds support for external entry points in rel and lib files when they are not provided in the ASM files. When the entry point is located in these external files, the tool will include omit exclusion of all outgoing references from the module containing the entry point. Additionally, if these references are functions, their traversed calls are also omitted from exclusion.

This is particularly useful in scenarios such as using the Arduino/Sduino library, where the main function is in the library and references `_setup()` and `_loop()`. With this update, the tool can correctly identify and retain `_setup()`, `_loop()`, and their related calls.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>